refactored nixgen to mkarray
