+++ nix_mx.cc
@@ -9,7 +9,7 @@
 #include "handle.h"
 #include "arguments.h"
 #include "struct.h"
-#include "nix2mx.h"
+#include "mknix.h"
 
 #include "nixfile.h"
 #include "nixsection.h"
