+++ src/nixdimensions.cc
@@ -1,11 +1,5 @@
-#include "nixtag.h"
-#include "nixgen.h"
-
-#include "mex.h"
-
 #include <nix.hpp>
-
-#include "handle.h"
+#include "mex.h"
 #include "arguments.h"
 #include "struct.h"
