+++ src/nixtag.cc
@@ -1,5 +1,5 @@
 #include "nixtag.h"
-#include "nixgen.h"
+#include "mkarray.h"
 
 #include "mex.h"
 
@@ -50,7 +50,7 @@ namespace nixtag {
         nix::Tag currObj = input.entity<nix::Tag>(1);
         double index = input.num<double>(2);
 
-        mxArray *data = nixgen::dataset_read_all(currObj.retrieveData(index));
+        mxArray *data = make_mx_array(currObj.retrieveData(index));
         output.set(0, data);
     }
 
@@ -58,7 +58,7 @@ namespace nixtag {
         nix::Tag currObj = input.entity<nix::Tag>(1);
         double index = input.num<double>(2);
 
-        mxArray *data = nixgen::dataset_read_all(currObj.retrieveFeatureData(index));
+        mxArray *data = make_mx_array(currObj.retrieveFeatureData(index));
         output.set(0, data);
     }
