+++ src/utils/mkarray.cc
@@ -1,4 +1,35 @@
 #include "mkarray.h"
+#include "mex.h"
+#include <nix.hpp>
+
+mxArray* make_mx_array(const nix::DataSet &da) {
+    nix::NDSize size = da.dataExtent();
+    const size_t len = size.size();
+    std::vector<mwSize> dims(len);
+
+    //NB: matlab is column-major, while HDF5 is row-major
+    //    data is correct with this, but dimensions don't
+    //    agree with the file anymore. Transpose it in matlab
+    //    (DataArray.read_all)
+    for (size_t i = 0; i < len; i++) {
+        dims[len - (i + 1)] = static_cast<mwSize>(size[i]);
+    }
+
+    nix::DataType da_type = da.dataType();
+    DType2 dtype = dtype_nix2mex(da_type);
+
+    if (!dtype.is_valid) {
+        throw std::domain_error("Unsupported data type");
+    }
+
+    mxArray *data = mxCreateNumericArray(dims.size(), dims.data(), dtype.cid, dtype.clx);
+    double *ptr = mxGetPr(data);
+
+    nix::NDSize offset(size.size(), 0);
+    da.getData(da_type, ptr, size, offset);
+
+    return data;
+}
 
 mxArray* make_mx_array(const nix::NDSize &size)
 {
